update card carousel styles
