+++ client/modules/Compare/components/Carousel/components/CardList/components/CardAssets/CardAssets.js
