+++ client/modules/Compare/components/Carousel/components/CardList/components/CardAssets/components/FavIcon/FavIcon.js
