+++ client/modules/Compare/components/Carousel/components/CardList/components/CardAssets/components/Image/Image.js
