+++ client/modules/Compare/components/Carousel/components/CardList/components/CardDetails/CardDetails.js
