+++ client/modules/Compare/components/Carousel/components/CardList/components/CardDetails/components/AvgRating/AvgRating.js
