+++ client/modules/Compare/components/Carousel/components/CardList/components/CardDetails/components/Category/Category.js
