+++ client/modules/Compare/components/Carousel/components/CardList/components/CardDetails/components/Category/__tests__/Category.js
