+++ client/modules/Compare/components/Carousel/components/CardList/components/CardDetails/components/Price/Price.js
