+++ client/modules/Compare/components/Carousel/components/CardList/components/CardDetails/components/Price/__tests__/Price.js
