+++ client/modules/Compare/components/Carousel/components/CardList/components/CardDetails/components/ProductName/ProductName.js
