+++ client/modules/Compare/components/Carousel/components/CardList/components/CardDetails/components/ProductName/__tests__/ProductName.js
