+++ client/modules/Compare/components/Carousel/components/RightButton/RightButton.js
@@ -1,12 +1,12 @@
 import React from 'react';
 
 const RightButton = () => {
-  const handleOnClick = () => {
+  const nextCard = () => {
     console.log('Right button was clicked');
   };
 
   return (
-    <button className="right-prev-button" type="button" lable="right" onClick={handleOnClick}>
+    <button className="right-prev-button" type="button" lable="right" onClick={nextCard}>
       <i className="fas fa-chevron-right" />
     </button>
   );
